@@ -5,89 +5,45 @@
   "author": "limin.zhang <limin.zhang@eveus.com>",
   "private": true,
   "scripts": {
-    "dev": "node build/dev-server.js",
-    "start": "node build/dev-server.js",
-    "build": "node build/build.js"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "serve": "vite preview"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.32",
-    "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "@fortawesome/vue-fontawesome": "^2.0.0",
-    "@types/axios": "^0.14.0",
-    "@types/lodash": "^4.14.112",
-    "@types/node": "^17.0.29",
-    "axios": "^0.16.2",
-    "bootstrap": "^4.5.3",
-    "bootstrap-vue": "^2.17.3",
-    "echarts": "^4.2.0-rc.2",
-    "element-ui": "^2.4.11",
-    "moment": "^2.22.2",
-    "v-charts": "^1.19.0",
-    "vue": "^2.5.16",
-    "vue-class-component": "^6.2.0",
-    "vue-i18n": "^8.0.0",
-    "vue-keyboard-multiselect": "^2.0.0",
-    "vue-property-decorator": "^9.1.2",
-    "vue-router": "^3.0.1",
-    "vue-socket.io": "^3.0.7",
-    "vue-types": "^1.5.3",
-    "vuex": "^3.0.1"
+    "@element-plus/icons-vue": "^2.3.1",
+    "@fortawesome/fontawesome-free": "^7.0.0",
+    "@fortawesome/fontawesome-svg-core": "^6.5.0",
+    "@fortawesome/free-solid-svg-icons": "^6.5.0",
+    "@fortawesome/vue-fontawesome": "^3.0.0",
+    "axios": "^1.6.0",
+    "bootstrap": "^5.3.0",
+    "echarts": "^5.4.0",
+    "element-plus": "^2.4.4",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.0",
+    "vue": "^3.4.0",
+    "vue-i18n": "^9.9.0",
+    "vue-router": "^4.2.0",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.6.0",
+    "@types/lodash": "^4.14.202",
     "@types/moment": "^2.13.0",
-    "autoprefixer": "^7.2.3",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-transform-decorators-legacy": "^1.3.5",
-    "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-preset-stage-1": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "babel-preset-stage-3": "^6.24.1",
-    "babel-register": "^6.22.0",
-    "babel-preset-es2015": "^6.24.1",
-    "chalk": "^2.4.1",
-    "connect-history-api-fallback": "^1.5.0",
-    "copy-webpack-plugin": "^4.5.2",
-    "css-loader": "^0.28.7",
-    "cssnano": "^3.10.0",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.16.3",
-    "extract-text-webpack-plugin": "^2.0.0",
-    "file-loader": "^0.11.2",
-    "friendly-errors-webpack-plugin": "^1.7.0",
-    "html-webpack-plugin": "^2.28.0",
-    "http-proxy-middleware": "^0.17.3",
-    "node-sass": "^4.9.2",
-    "opn": "^5.3.0",
-    "optimize-css-assets-webpack-plugin": "^2.0.0",
-    "ora": "^1.2.0",
-    "rimraf": "^2.6.0",
-    "sass-loader": "^6.0.6",
-    "semver": "^5.5.0",
-    "shelljs": "^0.7.6",
-    "style-loader": "^0.18.2",
-    "ts-loader": "^2.3.7",
-    "typescript": "^3.7.5",
-    "url-loader": "^0.5.8",
-    "vue-loader": "^15.2.4",
-    "vue-style-loader": "^4.1.0",
-    "vue-template-compiler": "^2.5.16",
-    "webpack": "^2.6.1",
-    "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-dev-middleware": "^1.12.2",
-    "webpack-hot-middleware": "^2.22.2",
-    "webpack-merge": "^4.1.3"
+    "@types/node": "^20.10.0",
+    "@types/vue-router": "^2.0.0",
+    "@vitejs/plugin-vue": "^4.5.0",
+    "sass": "^1.69.0",
+    "typescript": "^5.3.0",
+    "vite": "^5.0.0"
   },
   "engines": {
-    "node": ">= 4.0.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 16.0.0",
+    "npm": ">= 8.0.0"
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not ie <= 8"
+    "not dead"
   ]
 }
