@@ -1,44 +1,36 @@
 <template>
-    <div id="app">
-        <router-view></router-view>
-    </div>
+  <div id="app">
+    <router-view />
+  </div>
 </template>
 
-<script lang='ts'>
-import Vue, { ComponentOptions } from 'vue';
-import Component from 'vue-class-component';
-
-@Component
-export default class App extends Vue {
-    name: 'app'
-};
+<script setup lang="ts">
+// No additional logic needed in setup
 </script>
 
 <style>
 body {
-    margin: 0px;
-    padding: 0px;
-    background: #1F2D3D;
-    font-family: Helvetica Neue, Helvetica, PingFang SC, Hiragino Sans GB, Microsoft YaHei, SimSun, sans-serif;
-    font-size: 14px;
-    -webkit-font-smoothing: antialiased
+  margin: 0px;
+  padding: 0px;
+  background: #1F2D3D;
+  font-family: Helvetica Neue, Helvetica, PingFang SC, Hiragino Sans GB, Microsoft YaHei, SimSun, sans-serif;
+  font-size: 14px;
+  -webkit-font-smoothing: antialiased;
 }
 
 .appnav {
-    padding: 10px 0px 2px 0px;
-    margin-bottom: 5px;
-    border-bottom: 1px solid #EEEEEE;
+  padding: 10px 0px 2px 0px;
+  margin-bottom: 5px;
+  border-bottom: 1px solid #EEEEEE;
 }
 
 .appheader {
-    padding-left: 50px;
+  padding-left: 50px;
 }
 
 .appfooter-language {
-    text-align: center;
-    color: darkgray;
-    /*margin-top: 40px;*/
-    /*padding: 10px;*/
+  text-align: center;
+  color: darkgray;
 }
 
 .appfooter {
@@ -49,20 +41,22 @@ body {
 }
 
 .description {
-    color: darkgray;
+  color: darkgray;
 }
 
 #app {
-    margin: 0px;
-    padding: 0px;
-    width: 100%;
-    height: 100vh;
+  margin: 0px;
+  padding: 0px;
+  width: 100%;
+  height: 100vh;
 }
+
 .pagination {
-    margin-top: 14px;
-    text-align: right;
-    padding-right: 25px;
+  margin-top: 14px;
+  text-align: right;
+  padding-right: 25px;
 }
+
 .page-table {
   background-color: white;
   padding-bottom: 10px;
