@@ -5,15 +5,15 @@
       <header>{{ $t('Search') }}</header>
       <el-form :inline="true" :model="queryForm" class="demo-form-inline" label-width="170px">
         <el-form-item :label="$t('role.roleName')">
-          <el-input v-model="queryForm.roleName" :placeholder="$t('role.roleName')" size="mini"></el-input>
+          <el-input v-model="queryForm.roleName" :placeholder="$t('role.roleName')" size="small"></el-input>
         </el-form-item>
         <el-form-item align="right" style="margin-left: 220px">
-          <el-button type="primary" @click="handleAdd" :disabled="isButtonEnabled('system:role:insert')" size="mini">
-            <i class="el-icon-circle-plus-outline"/><span>{{ $t('option.add') }}</span>
+          <el-button type="primary" @click="handleAdd" :disabled="isButtonEnabled('system:role:insert')" size="small">
+            <el-icon><Plus /></el-icon><span>{{ $t('option.add') }}</span>
           </el-button>
           <el-button type="primary" @click="handleSynchronizePermission"
-                     :disabled="isButtonEnabled('system:role:insert')" size="mini">
-            <i class="el-icon-circle-plus-outline"/><span>{{ $t('option.syn') }}</span>
+                     :disabled="isButtonEnabled('system:role:insert')" size="small">
+            <el-icon><Plus /></el-icon><span>{{ $t('option.syn') }}</span>
           </el-button>
         </el-form-item>
       </el-form>
@@ -25,7 +25,8 @@
                   style="width: 97%; margin-top: 10px; margin-left: 20px;"
                   :row-class-name="tableRowClassName"
                   v-loading="ui.loading"
-                  align="center" border>
+                  align="center" 
+                  border>
           <el-table-column type="index" prop="id" :label="$t('common.index')" width="80"></el-table-column>
           <el-table-column prop="roleName" :label="$t('role.roleName')" min-width="80"></el-table-column>
           <el-table-column prop="description" :label="$t('role.description')"></el-table-column>
@@ -34,19 +35,19 @@
           <el-table-column prop="updateTime" :label="$t('common.updateTime')" width="230"
                            :formatter="defaultTimeFormatter"></el-table-column>
           <el-table-column :label="$t('common.option')" width="350">
-            <template slot-scope="scope">
-              <el-button size="mini" type="primary" class="normal-btn btn-bluelight" @click="handleEdit(scope.row)"
-                         :disabled="isButtonEnabledByUser(scope.row,'system:role:update')">{{ _self.$t('option.update') }}
+            <template #default="{ row }">
+              <el-button size="small" type="primary" class="normal-btn btn-bluelight" @click="handleEdit(row)"
+                         :disabled="isButtonEnabledByUser(row,'system:role:update')">{{ $t('option.update') }}
               </el-button>
-              <el-button size="mini" type="primary" class="normal-btn btn-greenlight"
-                         @click="handleAuthorize(scope.row)"
-                         :disabled="isButtonEnabledByUser(scope.row,'system:role:assign-permission')">
-                {{ _self.$t('common.authorize') }}
+              <el-button size="small" type="primary" class="normal-btn btn-greenlight"
+                         @click="handleAuthorize(row)"
+                         :disabled="isButtonEnabledByUser(row,'system:role:assign-permission')">
+                {{ $t('common.authorize') }}
               </el-button>
-              <el-button size="mini" type="primary" class="normal-btn btn-red"
-                         @click="handleDelete(scope.row)"
-                         :disabled="isButtonEnabledByUser(scope.row,'system:role:delete')">{{
-                  _self.$t('option.delete')
+              <el-button size="small" type="primary" class="normal-btn btn-red"
+                         @click="handleDelete(row)"
+                         :disabled="isButtonEnabledByUser(row,'system:role:delete')">{{
+                  $t('option.delete')
                 }}
               </el-button>
             </template>
@@ -55,29 +56,31 @@
       </el-row>
     </div>
 
-    <el-dialog :title="ui.addRecord ? $t('role.add') : $t('role.update')" :visible.sync="ui.dialogVisible" width="40%"
+    <el-dialog :title="ui.addRecord ? $t('role.add') : $t('role.update')" v-model="ui.dialogVisible" width="40%"
                class="addDio">
-      <el-form :model="editForm" :rules="editFormRules" label-width="100px" style="padding-right: 50px" ref="editForm1">
+      <el-form :model="editForm" :rules="editFormRules" label-width="100px" style="padding-right: 50px" ref="editFormRef">
         <el-form-item :label="$t('role.roleName')" prop="roleName">
-          <el-input v-model="editForm.roleName" autofocus size="mini"></el-input>
+          <el-input v-model="editForm.roleName" autofocus size="small"></el-input>
         </el-form-item>
         <el-form-item :label="$t('role.description')" prop="description">
-          <el-input v-model="editForm.description" size="mini"></el-input>
+          <el-input v-model="editForm.description" size="small"></el-input>
         </el-form-item>
       </el-form>
-      <div slot="footer" class="dialog-footer">
-        <el-button @click.native="ui.dialogVisible = false">{{ $t('common.cancel') }}</el-button>
-        <el-button type="primary" @click.native="handleSaveOrUpdate()">{{ $t('common.ok') }}</el-button>
-      </div>
+      <template #footer>
+        <div class="dialog-footer">
+          <el-button @click="ui.dialogVisible = false">{{ $t('common.cancel') }}</el-button>
+          <el-button type="primary" @click="handleSaveOrUpdate()">{{ $t('common.ok') }}</el-button>
+        </div>
+      </template>
     </el-dialog>
 
-    <el-dialog :title="$t('common.authorize')" :visible.sync="ui.permsDialogVisible" @open="loadPermsDialogData"
+    <el-dialog :title="$t('common.authorize')" v-model="ui.permsDialogVisible" @open="loadPermsDialogData"
                class="addDio">
       <el-row :gutter="20">
         <el-col :push="15" style="margin-bottom: 10px">
-          <el-button size="mini" @click="selectAll()" class="button-no-focus">{{ $t('common.select-all') }}
+          <el-button size="small" @click="selectAll()" class="button-no-focus">{{ $t('common.select-all') }}
           </el-button>
-          <el-button size="mini" @click="selectNone()" class="button-no-focus">{{ $t('common.select-none') }}
+          <el-button size="small" @click="selectNone()" class="button-no-focus">{{ $t('common.select-none') }}
           </el-button>
         </el-col>
       </el-row>
@@ -88,343 +91,271 @@
               show-checkbox
               default-expand-all
               node-key="tag"
-              ref="tree"
+              ref="treeRef"
               :props="defaultTreeProps"
               :default-checked-keys="['index','session:all']"
           >
           </el-tree>
         </div>
       </el-row>
-      <div slot="footer" class="dialog-footer">
-        <el-button @click.native="ui.permsDialogVisible = false">{{ $t('common.cancel') }}</el-button>
-        <el-button type="primary" @click.native="handlePermsUpdate()">{{ $t('common.ok') }}</el-button>
-      </div>
+      <template #footer>
+        <div class="dialog-footer">
+          <el-button @click="ui.permsDialogVisible = false">{{ $t('common.cancel') }}</el-button>
+          <el-button type="primary" @click="handlePermsUpdate()">{{ $t('common.ok') }}</el-button>
+        </div>
+      </template>
     </el-dialog>
-
   </div>
 </template>
-<script lang='ts'>
-import Vue from 'vue';
-import Component from 'vue-class-component';
-import i18n from '../i18n';
-import * as Utils from '../../utils';
-import * as utils from '../../utils';
-import * as nav from '../../common/nav';
 
-@Component
-export default class Roles extends Vue {
-  ui = {
-    loading: false,
-    dialogVisible: false,
-    addRecord: false,
-    permsDialogVisible: false,
-    authentication: false
-  };
-  tableData = [];
-  queryForm = {roleName: ''};
-  editForm = {id: '', roleName: '', description: ''};
-  editFormRules = {
-    roleName: [{required: true, message: i18n.t('role.inputRoleName'), trigger: 'blur'}],
-  };
-  disablePermissions: any = [];
-  authentication = {password: ''};
-  defaultTreeProps = {label: 'name', children: 'children'};
-  allPermissions = [];
-  permsEditForm = {roleId: '', oldPerms: [''], selectedPermissions: ['']};
-  active: number
-  row: any
-  Permissions = [] as any;
+<script setup lang="ts">
+import { ref, reactive, computed, onMounted } from 'vue'
+import { useStore } from 'vuex'
+import { useI18n } from 'vue-i18n'
+import { ElMessage, ElMessageBox, type FormInstance } from 'element-plus'
+import { Plus } from '@element-plus/icons-vue'
+import * as Utils from '../../utils'
+import * as nav from '../../common/nav'
 
-  get filteredTableData() {
-    let self = this;
-    var roleName = this.queryForm.roleName;
-    if (roleName == '') return self.tableData;
-    return self.tableData.filter(function (item: any) {
-      return item.roleName.indexOf(roleName) != -1;
-    });
-  };
+const { t } = useI18n()
+const store = useStore()
 
-  tableRowClassName(rowIndex: any) {
-    if (rowIndex.rowIndex % 2 === 0)
-      return '';
-    return 'warning-row';
+// Refs
+const editFormRef = ref<FormInstance>()
+const treeRef = ref()
+
+// Reactive data
+const ui = reactive({
+  loading: false,
+  dialogVisible: false,
+  addRecord: false,
+  permsDialogVisible: false,
+  authentication: false
+})
+
+const tableData = ref<any[]>([])
+const queryForm = reactive({ roleName: '' })
+const editForm = reactive({ id: '', roleName: '', description: '' })
+const editFormRules = reactive({
+  roleName: [{ required: true, message: t('role.inputRoleName'), trigger: 'blur' }],
+})
+
+const disablePermissions = ref<any[]>([])
+const authentication = reactive({ password: '' })
+const defaultTreeProps = { label: 'name', children: 'children' }
+const allPermissions = ref<any[]>([])
+const permsEditForm = reactive({ roleId: '', oldPerms: [''], selectedPermissions: [''] })
+const active = ref<number>(0)
+const row = ref<any>(null)
+const Permissions = ref<any[]>([])
+
+// Computed
+const filteredTableData = computed(() => {
+  const roleName = queryForm.roleName
+  if (roleName === '') return tableData.value
+  return tableData.value.filter((item: any) => {
+    return item.roleName.indexOf(roleName) !== -1
+  })
+})
+
+// Methods
+const tableRowClassName = ({ rowIndex }: { rowIndex: number }) => {
+  if (rowIndex % 2 === 0) return ''
+  return 'warning-row'
+}
+
+const isButtonEnabledByUser = (row: any, buttonName: string) => {
+  if (!row.removable && buttonName === 'system:role:delete') {
+    return true
   }
-
-  async handleAuthentication(row: any, active: number) {
-    if (active == 2) {
-      let confirmed = await Utils.confirm(this, String(i18n.t('common.deleteConfirm')), String(i18n.t('common.confirm')));
-      if (!confirmed)
-        return;
-    }
-    await utils.clearValidateForm(this.$refs.authentication)
-    this.ui.authentication = true
-    this.authentication.password = ''
-    this.row = row;
-    this.active = active;
+  if (!row.removable && buttonName === 'system:role:assign-permission') {
+    return true
   }
-
-  async mounted() {
-    await this.loadTableData();
-    // await this.getPermission();
-    // await this.handleSynchronizePermission();
-  };
-
-  isButtonEnabledByUser(row: any, buttonName: string) {
-    if (!row.removable && buttonName == 'system:role:delete') {
-      return true;
-    }
-    if (!row.removable && buttonName == 'system:role:assign-permission') {
-      return true;
-    }
-    if (!row.removable && buttonName == 'system:role:update') {
-      return true;
-    }
-    var state = this.$store.state.buttons.has(buttonName);
-    if (state)
-      return false;
-    return true;
-  };
-
-  isButtonEnabled(buttonName: string) {
-    var state = this.$store.state.buttons.has(buttonName);
-    if (state)
-      return false;
-    return true;
-  };
-
-
-  selectAll() {
-    // let permissions = nav.getPermissionTree();
-    // (this.$refs.tree as any).setCheckedNodes(permissions);
-    let permissions = nav.getPermissions();
-    let data = [] as any;
-    for (let value of permissions) {
-      data.push(value.tag)
-    }
-    (this.$refs.tree as any).setCheckedKeys(data);
+  if (!row.removable && buttonName === 'system:role:update') {
+    return true
   }
+  const state = store.state.buttons.has(buttonName)
+  return !state
+}
 
-  selectNone() {
-    (this.$refs.tree as any).setCheckedKeys(['session:all']);
+const isButtonEnabled = (buttonName: string) => {
+  const state = store.state.buttons.has(buttonName)
+  return !state
+}
+
+const selectAll = () => {
+  const permissions = nav.getPermissions()
+  const data: any[] = []
+  for (const value of permissions) {
+    data.push(value.tag)
   }
+  treeRef.value?.setCheckedKeys(data)
+}
 
+const selectNone = () => {
+  treeRef.value?.setCheckedKeys(['session:all'])
+}
 
-  async loadTableData() {
-    this.ui.loading = true;
-    let result = await Utils.doGet(this, '/api/admin/roles');
-    if (result.success) {
-      this.tableData = result.data.records;
+const loadTableData = async () => {
+  ui.loading = true
+  const result = await Utils.doGet({ $router: null }, '/api/admin/roles')
+  if (result.success) {
+    tableData.value = result.data.records
+  } else {
+    ElMessage.warning(t('common.loadFail'))
+  }
+  ui.loading = false
+}
+
+const defaultTimeFormatter = (row: any, column: any) => {
+  const source = row[column.property]
+  return Utils.formatDateString(source)
+}
+
+const handleAdd = async () => {
+  Utils.clearValidateForm(editFormRef.value)
+  ui.dialogVisible = true
+  ui.addRecord = true
+  editForm.roleName = ''
+  editForm.description = ''
+  editForm.id = ''
+}
+
+const handleEdit = async (rowData: any) => {
+  Utils.clearValidateForm(editFormRef.value)
+  ui.dialogVisible = true
+  ui.addRecord = false
+  editForm.roleName = rowData.roleName
+  editForm.description = rowData.description
+  editForm.id = rowData.id
+}
+
+const handleDelete = async (rowData: any) => {
+  try {
+    await ElMessageBox.confirm(t('common.deleteConfirm'), t('common.confirm'), { type: 'warning' })
+    const roleId = rowData.id
+    const result = await Utils.doDelete({ $router: null }, '/api/admin/roles/' + roleId, {})
+    if (!result.success) {
+      ElMessage.warning(t('common.deleteFail') + t(result.message))
     } else {
-      Utils.showWarning(String(i18n.t('common.loadFail')));
+      ElMessage.success(t('common.deleteSuccess'))
+      loadTableData()
     }
-    this.ui.loading = false;
-  };
+  } catch {
+    // User cancelled
+  }
+}
 
-  defaultTimeFormatter(row: any, column: any) {
-    var source = row[column.property];
-    return Utils.formatDateString(source);
-  };
-
-  async handleAdd() {
-    Utils.clearValidateForm(this.$refs.editForm1);
-    this.ui.dialogVisible = true;
-    this.ui.addRecord = true;
-    this.editForm.roleName = '';
-    this.editForm.description = '';
-    this.editForm.id = '';
-  };
-
-  async handleEdit(row: any) {
-    Utils.clearValidateForm(this.$refs.editForm1);
-    this.ui.dialogVisible = true;
-    this.ui.addRecord = false;
-    this.editForm.roleName = row.roleName;
-    this.editForm.description = row.description;
-    this.editForm.id = row.id;
-  };
-
-  async handleDelete(row: any) {
-    let confirmed = await Utils.confirm(this, String(i18n.t('common.deleteConfirm')), String(i18n.t('common.confirm')));
-    if (!confirmed) return;
-    var roleId = row.id;
-    var result = await Utils.doDelete(this, '/api/admin/roles/' + roleId, {});
-    if (!result.success)
-      Utils.showWarning(String(i18n.t('common.deleteFail')) + String(i18n.t(result.message)));
-    else {
-      Utils.showSuccess(String(i18n.t('common.deleteSuccess')));
-      this.loadTableData();
-    }
-  };
-
-  async handleSaveOrUpdate() {
-    let validated = await Utils.validateForm(this.$refs.editForm1);
-    if (!validated) return;
-    if (this.ui.addRecord) {
-      let record = this.editForm;
-      let result = await Utils.doPost(this, '/api/admin/roles', record);
+const handleSaveOrUpdate = async () => {
+  try {
+    await editFormRef.value?.validate()
+    if (ui.addRecord) {
+      const record = editForm
+      const result = await Utils.doPost({ $router: null }, '/api/admin/roles', record)
       if (!result.success) {
-        Utils.showWarning(String(i18n.t('common.insertFail')) + String(i18n.t(result.message)));
+        ElMessage.warning(t('common.insertFail') + t(result.message))
       } else {
-        Utils.showSuccess(String(i18n.t('common.insertSuccess')));
-        this.ui.dialogVisible = false;
-        this.loadTableData();
-      }
-    } else {
-      let record = this.editForm;
-      let result = await Utils.doPut(this, '/api/admin/roles/' + record.id, record);
-      if (!result.success) {
-        Utils.showWarning(String(i18n.t('common.updateFail')) + String(i18n.t(result.message)));
-      } else {
-        Utils.showSuccess(String(i18n.t('common.updateSuccess')));
-        this.ui.dialogVisible = false;
-        this.loadTableData();
+        ElMessage.success(t('common.updateSuccess'))
+        ui.dialogVisible = false
+        loadTableData()
       }
     }
-  };
+  } catch {
+    // Validation failed
+  }
+}
 
-  async handleAuthorize(row: any) {
-    this.permsEditForm.roleId = row.id;
-    this.allPermissions = [];
-    this.ui.permsDialogVisible = true;
-    this.authentication.password = '';
-  };
+const handleAuthorize = async (rowData: any) => {
+  permsEditForm.roleId = rowData.id
+  allPermissions.value = []
+  ui.permsDialogVisible = true
+  authentication.password = ''
+}
 
-  async loadPermsDialogData() {
-    let permissins: any = nav.getPermissionTree()
-    this.allPermissions = permissins
-    let selectedPermissions = await this.loadAllPermsByRoleId(this.permsEditForm.roleId);
-    (this.$refs.tree as any).setCheckedKeys(selectedPermissions);
-  };
+const loadPermsDialogData = async () => {
+  const permissions: any = nav.getPermissionTree()
+  allPermissions.value = permissions
+  const selectedPermissions = await loadAllPermsByRoleId(permsEditForm.roleId)
+  treeRef.value?.setCheckedKeys(selectedPermissions)
+}
 
-  // async getPermission() {
-  //   let permissins: any = nav.getPermissionTree()
-  //   for (let i in permissins) {
-  //     await this.addPermission(permissins[i])
-  //   }
-  //   // console.log(this.Permissions)
-  // }
-  //
-  // async addPermission(permissins: any) {
-  //   let permissinMenu = {} as any
-  //   permissinMenu.tag = permissins.tag
-  //   permissinMenu.type = permissins.type
-  //   if (permissins.pattern !== undefined)
-  //     permissinMenu.pattern = permissins.pattern
-  //   permissinMenu.whiteList = permissins.whiteList
-  //   if (permissins.method !== undefined)
-  //     permissinMenu.method = permissins.method
-  //   this.Permissions.push(permissinMenu)
-  //   for (let i in permissins.children) {
-  //     await this.addPermission(permissins.children[i])
-  //   }
-  // }
-
-  async handlePermsUpdate() {
-    let keys = (this.$refs.tree as any).getCheckedKeys().concat((this.$refs.tree as any).getHalfCheckedKeys());
-    let oldPerms = this.permsEditForm.oldPerms;
-    if (keys.length == oldPerms.length) {
-      var oldPermsSet = new Set(oldPerms);
-      var diffArr = keys.filter((item: any) => {
-        return !oldPermsSet.has(item);
-      });
-      if (diffArr.length == 0) {
-        this.ui.permsDialogVisible = false;
-        Utils.showInfo(String(i18n.t('common.noChange')));
-        return;
-      }
-    }
-    this.permsEditForm.selectedPermissions = keys;
-    var result = await Utils.doPut(this, '/api/admin/roleperms/' + this.permsEditForm.roleId, this.permsEditForm);
-    if (result.success) {
-      Utils.showSuccess(String(i18n.t('common.updateSuccess')));
-      this.ui.permsDialogVisible = false;
-    } else {
-      Utils.showWarning(String(i18n.t('common.updateFail')) + i18n.t(result.message));
+const handlePermsUpdate = async () => {
+  const keys = treeRef.value?.getCheckedKeys().concat(treeRef.value?.getHalfCheckedKeys())
+  const oldPerms = permsEditForm.oldPerms
+  if (keys.length === oldPerms.length) {
+    const oldPermsSet = new Set(oldPerms)
+    const diffArr = keys.filter((item: any) => {
+      return !oldPermsSet.has(item)
+    })
+    if (diffArr.length === 0) {
+      ui.permsDialogVisible = false
+      ElMessage.info(t('common.noChange'))
+      return
     }
   }
+  permsEditForm.selectedPermissions = keys
+  const result = await Utils.doPut({ $router: null }, '/api/admin/roleperms/' + permsEditForm.roleId, permsEditForm)
+  if (result.success) {
+    ElMessage.success(t('common.updateSuccess'))
+    ui.permsDialogVisible = false
+  } else {
+    ElMessage.warning(t('common.updateFail') + t(result.message))
+  }
+}
 
-  async handleSynchronizePermission() {
-    let result = await Utils.doPost(this, '/api/admin/permissions', nav.getPermissions());
-    if (result.success) {
-      Utils.showSuccess(String(i18n.t('common.updateSuccess')));
-    } else {
-      Utils.showWarning(String(i18n.t('common.updateFail')) + i18n.t(result.message));
+const handleSynchronizePermission = async () => {
+  const result = await Utils.doPost({ $router: null }, '/api/admin/permissions', nav.getPermissions())
+  if (result.success) {
+    ElMessage.success(t('common.updateSuccess'))
+  } else {
+    ElMessage.warning(t('common.updateFail') + t(result.message))
+  }
+}
+
+const loadAllPermsByRoleId = async (roleId: any) => {
+  const result = await Utils.doGet({ $router: null }, '/api/admin/roleperms/' + roleId)
+  if (result.success) {
+    const oldPerms = buildOldPerms(result.data)
+    permsEditForm.oldPerms = oldPerms
+    disablePermissions.value = buildDisabledPerms(result.data)
+    return oldPerms
+  } else {
+    ElMessage.warning(t('common.loadFail') + t(result.message))
+  }
+  return []
+}
+
+const buildOldPerms = (data: any) => {
+  const perms = []
+  if (data != null) {
+    for (let i = 0; i < data.length; i++) {
+      perms.push(data[i].tag)
     }
   }
+  if (perms.length === 0) {
+    perms.push("session:all")
+  }
+  return perms
+}
 
-  async loadAllPermsByRoleId(roleId: any) {
-    let result = await Utils.doGet(this, '/api/admin/roleperms/' + roleId);
-    if (result.success) {
-      let oldPerms = this.buildOldPerms(result.data);
-      // console.log(oldPerms)
-      this.permsEditForm.oldPerms = oldPerms;
-      this.disablePermissions = this.buildDisabledPerms(result.data);
-      return oldPerms;
-    } else {
-      Utils.showWarning(String(i18n.t('common.loadFail')) + i18n.t(result.message));
-    }
-    return [];
-  };
-
-  buildOldPerms(data: any) {
-    let perms = [];
-    if (data != null) {
-      for (let i = 0; i < data.length; i++) {
-        perms.push(data[i].tag);
+const buildDisabledPerms = (data: any) => {
+  const perms = []
+  if (data != null) {
+    for (let i = 0; i < data.length; i++) {
+      if (data[i].removable === 0) {
+        perms.push(data[i].tag)
       }
     }
-    if (perms.length === 0) {
-      perms.push("session:all");
-    }
-    return perms;
   }
+  return perms
+}
 
-  buildDisabledPerms(data: any) {
-    let perms = [];
-    if (data != null) {
-      for (let i = 0; i < data.length; i++) {
-        if (data[i].removable == 0) {
-          perms.push(data[i].tag);
-        }
-      }
-    }
-    return perms;
-  }
-
-  // toggleTreeNodeCheckbox(permissions: any, disablePermission: any) {
-  //   let length = permissions.length;
-  //   for (let i = 0; i < length; i++) {
-  //     let permission = permissions[i];
-  //     if (disablePermission.includes(permission.tag)) {
-  //       permission.disable = true;
-  //     }
-  //     if (permission.children != null) {
-  //       let children = permission.children;
-  //       let childrenLength = children.length;
-  //       for (let j = 0; j < childrenLength; j++) {
-  //         let child = children[j];
-  //         if (disablePermission.includes(child.tag)) {
-  //           console.log(child.tag)
-  //           child.disabled = true;
-  //         }
-  //         if (child.options != null) {
-  //           let options = child.options;
-  //           let optionsLength = options.length;
-  //           for (let k = 0; k < optionsLength; k++) {
-  //             let option = options[k];
-  //             if (disablePermission.includes(option.tag)) {
-  //               console.log(option.tag)
-  //               option.disabled = true;
-  //             }
-  //           }
-  //         }
-  //       }
-  //     }
-  //   }
-  // }
-
-};
+// Lifecycle
+onMounted(async () => {
+  await loadTableData()
+})
 </script>
+
 <style scoped lang="scss">
 .compact-form {
   margin: 0px;
