@@ -4,15 +4,16 @@
       <header>{{ $t('Search') }}</header>
       <el-form :inline="true" :model="queryForm" class="demo-form-inline" label-width="170px">
         <el-form-item :label="$t('user.userName')">
-          <el-input v-model="queryForm.userName" :placeholder="$t('user.userName')" size="mini"></el-input>
+          <el-input v-model="queryForm.userName" :placeholder="$t('user.userName')" size="small"></el-input>
         </el-form-item>
         <el-form-item align="right" style="margin-left: 220px">
-          <el-button type="primary" @click="handleAdd" :disabled="isButtonEnabled('system:user:insert')" size="mini">
-            <i class="el-icon-circle-plus-outline"/><span>{{ $t('option.add') }}</span>
+          <el-button type="primary" @click="handleAdd" :disabled="isButtonEnabled('system:user:insert')" size="small">
+            <el-icon><Plus /></el-icon><span>{{ $t('option.add') }}</span>
           </el-button>
         </el-form-item>
       </el-form>
     </div>
+    
     <div class="page-table" style="padding-bottom: 40px">
       <el-row :gutter="20">
         <el-table :data="filteredTableData" style="width: 97%;margin-top: 10px;margin-left: 20px;"
@@ -25,23 +26,23 @@
                            :formatter="defaultTimeFormatter"></el-table-column>
           <el-table-column prop="updateTime" :label="$t('common.updateTime')" width="180"
                            :formatter="defaultTimeFormatter"></el-table-column>
-          <el-table-column :label="$t('common.option')" :context="_self" width="480">
-            <template slot-scope="scope">
-              <el-button size="mini" type="primary" class="normal-btn btn-bluelight"
+          <el-table-column :label="$t('common.option')" width="480">
+            <template #default="{ row }">
+              <el-button size="small" type="primary" class="normal-btn btn-bluelight"
                          :disabled="isButtonEnabled('system:user:update')"
-                         @click="handleEdit(scope.row)">{{ _self.$t('option.update') }}
+                         @click="handleEdit(row)">{{ $t('option.update') }}
               </el-button>
-              <el-button size="mini" type="primary" class="normal-btn btn-greenlight"
-                         :disabled="isButtonEnabled('system:user:assign-roles')||scope.row.userName===userNameLogin || scope.row.removable!='1'"
-                         @click="handleSelectRoles(scope.row)">{{ _self.$t('user.allocateRole') }}
+              <el-button size="small" type="primary" class="normal-btn btn-greenlight"
+                         :disabled="isButtonEnabled('system:user:assign-roles')||row.userName===userNameLogin || row.removable!='1'"
+                         @click="handleSelectRoles(row)">{{ $t('user.allocateRole') }}
               </el-button>
-              <el-button size="mini" type="primary" class="normal-btn btn-red"
-                         :disabled="isButtonEnabled('system:user:reset-password')||scope.row.userName===userNameLogin"
-                         @click="resetPassword(scope.row)">{{ _self.$t('user.resetPassword') }}
+              <el-button size="small" type="primary" class="normal-btn btn-red"
+                         :disabled="isButtonEnabled('system:user:reset-password')||row.userName===userNameLogin"
+                         @click="resetPassword(row)">{{ $t('user.resetPassword') }}
               </el-button>
-              <el-button size="mini" type="primary" class="normal-btn btn-red"
-                         :disabled="isButtonEnabled('system:user:delete')||scope.row.userName===userNameLogin || scope.row.removable!='1'"
-                         @click="handleDelete(scope.row,3)">{{ _self.$t('option.delete') }}
+              <el-button size="small" type="primary" class="normal-btn btn-red"
+                         :disabled="isButtonEnabled('system:user:delete')||row.userName===userNameLogin || row.removable!='1'"
+                         @click="handleDelete(row)">{{ $t('option.delete') }}
               </el-button>
             </template>
           </el-table-column>
@@ -49,37 +50,41 @@
       </el-row>
     </div>
 
-    <el-dialog :title="ui.addRecord?$t('user.add'):$t('user.update')" :visible.sync="ui.dialogVisible" width="40%"
+    <el-dialog :title="ui.addRecord?$t('user.add'):$t('user.update')" v-model="ui.dialogVisible" width="40%"
                class="addDio">
-      <el-form :model="editForm" :rules="editFormRules" label-width="130px" ref="editForm" style="padding-right: 50px">
-        <input type="hidden" v-model="editForm.id"></input>
+      <el-form :model="editForm" :rules="editFormRules" label-width="130px" ref="editFormRef" style="padding-right: 50px">
+        <input type="hidden" v-model="editForm.id">
         <el-form-item :label="$t('user.userName')" prop="userName">
-          <el-input v-model="editForm.userName" autofocus :readonly="!ui.addRecord" size="mini"></el-input>
+          <el-input v-model="editForm.userName" autofocus :readonly="!ui.addRecord" size="small"></el-input>
         </el-form-item>
         <el-form-item :label="$t('user.realName')" prop="realName">
-          <el-input v-model="editForm.realName" size="mini"></el-input>
+          <el-input v-model="editForm.realName" size="small"></el-input>
         </el-form-item>
       </el-form>
-      <div slot="footer" class="dialog-footer">
-        <el-button @click.native="ui.dialogVisible = false">{{ $t('common.cancel') }}</el-button>
-        <el-button type="primary" @click.native="handleSaveOrUpdate()">{{ $t('common.ok') }}</el-button>
-      </div>
+      <template #footer>
+        <div class="dialog-footer">
+          <el-button @click="ui.dialogVisible = false">{{ $t('common.cancel') }}</el-button>
+          <el-button type="primary" @click="handleSaveOrUpdate()">{{ $t('common.ok') }}</el-button>
+        </div>
+      </template>
     </el-dialog>
 
-    <el-dialog :title="$t('ResetPassword')" :visible.sync="ui.resetPassword" width="40%" class="addDio">
-      <el-form :model="password" :rules="passwordRules" label-width="130px" ref="editForm" style="padding-right: 50px">
-        <input type="hidden" v-model="editForm.id"></input>
+    <el-dialog :title="$t('ResetPassword')" v-model="ui.resetPassword" width="40%" class="addDio">
+      <el-form :model="password" :rules="passwordRules" label-width="130px" ref="passwordFormRef" style="padding-right: 50px">
+        <input type="hidden" v-model="editForm.id">
         <el-form-item :label="$t('user.password')" prop="password">
-          <el-input v-model="password.password" size="mini"></el-input>
+          <el-input v-model="password.password" size="small"></el-input>
         </el-form-item>
       </el-form>
-      <div slot="footer" class="dialog-footer">
-        <el-button @click.native="ui.dialogVisible = false">{{ $t('common.cancel') }}</el-button>
-        <el-button type="primary" @click.native="handleResetPassword()">{{ $t('common.ok') }}</el-button>
-      </div>
+      <template #footer>
+        <div class="dialog-footer">
+          <el-button @click="ui.resetPassword = false">{{ $t('common.cancel') }}</el-button>
+          <el-button type="primary" @click="handleResetPassword()">{{ $t('common.ok') }}</el-button>
+        </div>
+      </template>
     </el-dialog>
 
-    <el-dialog :title="$t('user.selectRole')" width="35%" :visible.sync="ui.rolesDialogVisible"
+    <el-dialog :title="$t('user.selectRole')" width="35%" v-model="ui.rolesDialogVisible"
                @open="loadRolesDialogData"
                class="addDio">
       <el-tree
@@ -88,244 +93,255 @@
           default-expand-all
           node-key="id"
           :props="defaultTreeProps"
-          ref="tree"
+          ref="rolesTreeRef"
           hightlight-current>
       </el-tree>
-      <div slot="footer" class="dialog-footer">
-        <el-button @click.native="ui.rolesDialogVisible = false">{{ $t('common.cancel') }}</el-button>
-        <el-button type="primary" @click.native="handleRolesUpdate()">{{ $t('common.ok') }}</el-button>
-      </div>
+      <template #footer>
+        <div class="dialog-footer">
+          <el-button @click="ui.rolesDialogVisible = false">{{ $t('common.cancel') }}</el-button>
+          <el-button type="primary" @click="handleRolesUpdate()">{{ $t('common.ok') }}</el-button>
+        </div>
+      </template>
     </el-dialog>
   </div>
 </template>
-<script lang='ts'>
-import Vue from 'vue';
-import Component from 'vue-class-component';
-import * as Utils from '../../utils';
-import i18n from '../i18n';
-import * as utils from "../../utils";
 
-@Component
-export default class Users extends Vue {
-  ui = {
-    loading: false,
-    dialogVisible: false,
-    addRecord: false,
-    rolesDialogVisible: false,
-    authentication: false,
-    resetPassword: false
-  };
-  Provider: []
-  editForm = {id: '', userName: '', realName: ''};
-  editFormRules = {
-    userName: [{required: true, message: i18n.t("user.inputUserName"), trigger: 'blur'}],
-    realName: [{required: true, message: i18n.t("user.inputRealName"), trigger: 'blur'}]
-  };
-  passwordRules = {
-    password: [{required: true, message: i18n.t("user.inputUserName"), trigger: 'blur'}],
-  }
-  password = {password: '', id: ''}
-  userNameLogin = ''
-  queryForm = {userName: ''};
-  tableData = [];
-  allRoles = [];
-  defaultTreeProps = {label: 'roleName', children: 'children'};
-  rolesEditForm: any = {id: '', roles: [], userId: '', oldRoles: ['']};
-  authentication = {password: ''};
-  authenticationFormRules = {
-    password: [{
-      required: true,
-      message: i18n.t('login.pwdisnull'),
-      trigger: 'blur'
-    }],
-  };
-  active: number
-  row: any
-  isEyes: boolean = false;
+<script setup lang="ts">
+import { ref, reactive, computed, onMounted } from 'vue'
+import { useStore } from 'vuex'
+import { useI18n } from 'vue-i18n'
+import { ElMessage, ElMessageBox, type FormInstance } from 'element-plus'
+import { Plus } from '@element-plus/icons-vue'
+import * as Utils from '../../utils'
 
-  showPassword() {
-    this.isEyes = !this.isEyes;
-  }
+const { t } = useI18n()
+const store = useStore()
 
-  get filteredTableData() {
-    let self = this;
-    return self.tableData;
-  }
+// Refs
+const editFormRef = ref<FormInstance>()
+const passwordFormRef = ref<FormInstance>()
+const rolesTreeRef = ref()
 
-  async resetPassword(row: any) {
-    this.ui.resetPassword = true
-    this.password.password = ''
-    this.password.id = row.id;
-  }
+// Reactive data
+const ui = reactive({
+  loading: false,
+  dialogVisible: false,
+  addRecord: false,
+  rolesDialogVisible: false,
+  authentication: false,
+  resetPassword: false
+})
 
-  async mounted() {
-    await this.loadTableData();
-    let userString = sessionStorage.getItem("user");
-    if (userString) {
-      let user = JSON.parse(userString);
-      this.userNameLogin = user.userName;
-    }
-  };
+const Provider = ref<any[]>([])
+const editForm = reactive({ id: '', userName: '', realName: '' })
 
-  isButtonEnabled(buttonName: string) {
-    var state = this.$store.state.buttons.has(buttonName);
-    if (state)
-      return false;
-    return true;
-  };
+const editFormRules = reactive({
+  userName: [{ required: true, message: t("user.inputUserName"), trigger: 'blur' }],
+  realName: [{ required: true, message: t("user.inputRealName"), trigger: 'blur' }]
+})
 
-  tableRowClassName(rowIndex: any) {
-    if (rowIndex.rowIndex % 2 === 0)
-      return '';
-    return 'warning-row';
-  }
+const passwordRules = reactive({
+  password: [{ required: true, message: t("user.inputUserName"), trigger: 'blur' }],
+})
 
-  async handleAdd() {
-    Utils.clearValidateForm(this.$refs.editForm);
-    this.ui.dialogVisible = true;
-    this.ui.addRecord = true;
-    this.editForm.userName = '';
-    this.editForm.realName = '';
-    this.editForm.id = '';
-  };
+const password = reactive({ password: '', id: '' })
+const userNameLogin = ref('')
+const queryForm = reactive({ userName: '' })
+const tableData = ref<any[]>([])
+const allRoles = ref<any[]>([])
+const defaultTreeProps = { label: 'roleName', children: 'children' }
+const rolesEditForm = reactive({ id: '', roles: [], userId: '', oldRoles: [''] })
+const authentication = reactive({ password: '' })
+const active = ref<number>(0)
+const row = ref<any>(null)
+const isEyes = ref(false)
 
-  async handleEdit(row: any) {
-    Utils.clearValidateForm(this.$refs.editForm);
-    this.ui.dialogVisible = true;
-    this.ui.addRecord = false;
-    this.editForm.userName = row.userName;
-    this.editForm.realName = row.realName;
-    this.editForm.id = row.id;
-  };
+// Computed
+const filteredTableData = computed(() => {
+  return tableData.value
+})
 
-  async handleDelete(row: any) {
-    let confirmed = await Utils.confirm(this, String(i18n.t('common.deleteConfirm')), String(i18n.t('common.confirm')));
-    if (confirmed) {
-      var roleId = row.id;
-      var result = await Utils.doDelete(this, '/api/admin/users/' + roleId, {});
-      if (!result.success)
-        Utils.showWarning(String(i18n.t('common.deleteFail')) + String(i18n.t(result.message)));
-      else {
-        Utils.showSuccess(String(i18n.t('common.deleteSuccess')));
-        this.loadTableData();
-      }
+// Methods
+const showPassword = () => {
+  isEyes.value = !isEyes.value
+}
+
+const resetPassword = async (rowData: any) => {
+  ui.resetPassword = true
+  password.password = ''
+  password.id = rowData.id
+}
+
+const isButtonEnabled = (buttonName: string) => {
+  const state = store.state.buttons.has(buttonName)
+  return !state
+}
+
+const tableRowClassName = ({ rowIndex }: { rowIndex: number }) => {
+  if (rowIndex % 2 === 0) return ''
+  return 'warning-row'
+}
+
+const handleAdd = async () => {
+  Utils.clearValidateForm(editFormRef.value)
+  ui.dialogVisible = true
+  ui.addRecord = true
+  editForm.userName = ''
+  editForm.realName = ''
+  editForm.id = ''
+}
+
+const handleEdit = async (rowData: any) => {
+  Utils.clearValidateForm(editFormRef.value)
+  ui.dialogVisible = true
+  ui.addRecord = false
+  editForm.userName = rowData.userName
+  editForm.realName = rowData.realName
+  editForm.id = rowData.id
+}
+
+const handleDelete = async (rowData: any) => {
+  try {
+    await ElMessageBox.confirm(t('common.deleteConfirm'), t('common.confirm'), { type: 'warning' })
+    const roleId = rowData.id
+    const result = await Utils.doDelete({ $router: null }, '/api/admin/users/' + roleId, {})
+    if (!result.success) {
+      ElMessage.warning(t('common.deleteFail') + t(result.message))
     } else {
-      console.log("canceled.");
+      ElMessage.success(t('common.deleteSuccess'))
+      loadTableData()
     }
+  } catch {
+    console.log("canceled.")
+  }
+}
 
-  };
-
-  async handleSaveOrUpdate() {
-    let validated = await Utils.validateForm(this.$refs.editForm);
-    if (!validated) return;
-    if (this.ui.addRecord) {
-      let record = this.editForm;
-      let result = await Utils.doPost(this, '/api/admin/users', record);
+const handleSaveOrUpdate = async () => {
+  try {
+    await editFormRef.value?.validate()
+    if (ui.addRecord) {
+      const record = editForm
+      const result = await Utils.doPost({ $router: null }, '/api/admin/users', record)
       if (!result.success) {
-        Utils.showWarning(String(i18n.t('common.insertFail')) + String(i18n.t(result.message)));
+        ElMessage.warning(t('common.insertFail') + t(result.message))
       } else {
-        Utils.showSuccess(String(i18n.t('common.insertSuccess')));
-        this.ui.dialogVisible = false;
-        this.loadTableData();
+        ElMessage.success(t('common.insertSuccess'))
+        ui.dialogVisible = false
+        loadTableData()
       }
     } else {
-      let record = this.editForm;
-      let result = await Utils.doPut(this, '/api/admin/users/' + record.id, record);
+      const record = editForm
+      const result = await Utils.doPut({ $router: null }, '/api/admin/users/' + record.id, record)
       if (!result.success) {
-        Utils.showWarning(String(i18n.t('common.updateFail')) + String(i18n.t(result.message)));
+        ElMessage.warning(t('common.updateFail') + t(result.message))
       } else {
-        Utils.showSuccess(String(i18n.t('common.updateSuccess')));
-        this.ui.dialogVisible = false;
-        this.loadTableData();
+        ElMessage.success(t('common.updateSuccess'))
+        ui.dialogVisible = false
+        loadTableData()
       }
     }
-  };
+  } catch {
+    // Validation failed
+  }
+}
 
-  async loadTableData() {
-    let result = await Utils.doGet(this, '/api/admin/users');
+const loadTableData = async () => {
+  const result = await Utils.doGet({ $router: null }, '/api/admin/users')
+  if (result.success) {
+    tableData.value = result.data.records
+  } else {
+    ElMessage.warning(t('common.loadFail'))
+  }
+}
+
+const rolesFormatter = (row: any, column: any) => {
+  const roles = row.roles
+  const roleNames = roles.map((item: any) => {
+    return item.roleName
+  })
+  return roleNames.join(", ")
+}
+
+const defaultTimeFormatter = (row: any, column: any) => {
+  const source = row[column.property]
+  return Utils.formatDateString(source)
+}
+
+const handleSelectRoles = async (rowData: any) => {
+  rolesEditForm.userId = rowData.id
+  allRoles.value = []
+  const oldRows = rowData.roles.map((item: any) => {
+    return item.id
+  })
+  rolesEditForm.oldRoles = oldRows
+  ui.rolesDialogVisible = true
+}
+
+const handleResetPassword = async () => {
+  try {
+    await ElMessageBox.confirm(t('user.resetPasswordConfirm'), t('common.confirm'), { type: 'warning' })
+    const result = await Utils.doPost({ $router: null }, '/api/admin/users/password', password)
     if (result.success) {
-      this.tableData = result.data.records;
+      ElMessage.info(t('user.resetPasswordSuccess') + result.data + ']')
+      ui.resetPassword = false
     } else {
-      Utils.showWarning(String(i18n.t('common.loadFail')));
+      ElMessage.warning(t('user.resetPasswordFail') + t(result.message))
     }
-  };
+  } catch {
+    // User cancelled
+  }
+}
 
-  rolesFormatter(row: any, column: any) {
-    var roles = row.roles;
-    var roleNames = roles.map((item: any) => {
-      return item.roleName;
-    });
-    return roleNames.join(", ");
-  };
+const loadAllRoles = async () => {
+  const result = await Utils.doGet({ $router: null }, '/api/admin/roles')
+  if (result.success) {
+    allRoles.value = result.data.records
+  } else {
+    ElMessage.warning(t('common.loadFail') + t(result.message))
+  }
+}
 
-  defaultTimeFormatter(row: any, column: any) {
-    var source = row[column.property];
-    return Utils.formatDateString(source);
-  };
+const loadRolesDialogData = async () => {
+  await loadAllRoles()
+  rolesTreeRef.value?.setCheckedKeys(rolesEditForm.oldRoles)
+}
 
-  async handleSelectRoles(row: any) {
-    this.rolesEditForm.userId = row.id;
-    this.allRoles = [];
-    var oldRows = row.roles.map((item: any) => {
-      return item.id;
-    });
-    this.rolesEditForm.oldRoles = oldRows;
-    this.ui.rolesDialogVisible = true;
-  };
-
-  async handleResetPassword() {
-    let confirmed = await Utils.confirm(this, String(i18n.t('user.resetPasswordConfirm')), String(i18n.t('common.confirm')));
-    if (!confirmed)
-      return;
-    let result = await Utils.doPost(this, '/api/admin/users/password', this.password);
-    if (result.success) {
-      Utils.messageInfo(String(i18n.t('user.resetPasswordSuccess')) + result.data + ']');
-      this.ui.resetPassword = false;
-    } else {
-      Utils.showWarning(String(i18n.t('user.resetPasswordFail')) + String(i18n.t(result.message)));
+const handleRolesUpdate = async () => {
+  const keys = rolesTreeRef.value?.getCheckedKeys()
+  const oldRoles = rolesEditForm.oldRoles
+  if (keys.length === oldRoles.length) {
+    const oldRolesSet = new Set(oldRoles)
+    const diffArr = keys.filter((item: any) => {
+      return !oldRolesSet.has(item)
+    })
+    if (diffArr.length === 0) {
+      ui.rolesDialogVisible = false
+      ElMessage.info(t('common.noChange'))
+      return
     }
-  };
+  }
+  rolesEditForm.selectedRoles = keys
+  const result = await Utils.doPut({ $router: null }, '/api/admin/userroles/' + rolesEditForm.userId, rolesEditForm)
+  if (result.success) {
+    ElMessage.success(t('common.updateSuccess'))
+    loadTableData()
+    ui.rolesDialogVisible = false
+  } else {
+    ElMessage.warning(t('common.updateFail') + t(result.message))
+  }
+}
 
-  async loadAllRoles() {
-    let result = await Utils.doGet(this, '/api/admin/roles');
-    if (result.success) {
-      this.allRoles = result.data.records;
-    } else {
-      Utils.showWarning(String(i18n.t('common.loadFail')) + String(i18n.t(result.message)));
-    }
-  };
-
-  async loadRolesDialogData() {
-    await this.loadAllRoles();
-    (this.$refs.tree as any).setCheckedKeys(this.rolesEditForm.oldRoles);
-  };
-
-  async handleRolesUpdate() {
-    var keys = (this.$refs.tree as any).getCheckedKeys();
-    var oldRoles = this.rolesEditForm.oldRoles;
-    if (keys.length == oldRoles.length) {
-      var oldRolesSet = new Set(oldRoles);
-      var diffArr = keys.filter((item: any) => {
-        return !oldRolesSet.has(item);
-      });
-      if (diffArr.length == 0) {
-        this.ui.rolesDialogVisible = false;
-        Utils.showInfo(String(i18n.t('common.noChange')));
-        return;
-      }
-    }
-    this.rolesEditForm.selectedRoles = keys;
-    var result = await Utils.doPut(this, '/api/admin/userroles/' + this.rolesEditForm.userId, this.rolesEditForm);
-    if (result.success) {
-      Utils.showSuccess(String(i18n.t('common.updateSuccess')));
-      this.loadTableData();
-      this.ui.rolesDialogVisible = false;
-    } else {
-      Utils.showWarning(String(i18n.t('common.updateFail')) + String(i18n.t(result.message)));
-    }
-  };
-};
+// Lifecycle
+onMounted(async () => {
+  await loadTableData()
+  const userString = sessionStorage.getItem("user")
+  if (userString) {
+    const user = JSON.parse(userString)
+    userNameLogin.value = user.userName
+  }
+})
 </script>
+
 <style scoped lang="scss">
 .compact-form {
   margin: 0px;
@@ -339,7 +355,6 @@ export default class Users extends Vue {
 .danger {
   color: red;
 }
-
 
 .danger:disabled {
   color: #bfcbd9;
