@@ -5,21 +5,23 @@
         <el-col :span="19" class="logo" style="color: #115484">
           <img src="../../assets/mk-logo.png"
                style="height: 40px; width: 130px; padding-left: 50px;padding-right: 30px"/>
-          <span style="margin-top:5px;color: #115484;font-size: 26px;font-family: Roboto-Bold;">MK - Visitor Management
-					</span>
+          <span style="margin-top:5px;color: #115484;font-size: 26px;font-family: Roboto-Bold;">MK - Visitor Management</span>
         </el-col>
         <el-col :span="2" class="userinfo">
           <el-dropdown trigger="click">
-            <span class="el-dropdown-link userinfo-inner" style="color: #115484;font-size: 16px"><img
-                src="../../assets/user2.png" style="width: 35px; height: 35px;"> {{ sysUserName }}</span>
-            <el-dropdown-menu slot="dropdown">
-              <el-dropdown-item @click.native="handleChangePassword">
-                <i class="el-icon-fa-edit"/>{{ $t("home.funpwd") }}
-              </el-dropdown-item>
-              <el-dropdown-item divided @click.native="logout">
-                <i class="el-icon-fa-sign-out-alt" style="width: 15px"/>{{ $t("home.funlogout") }}
-              </el-dropdown-item>
-            </el-dropdown-menu>
+            <span class="el-dropdown-link userinfo-inner" style="color: #115484;font-size: 16px">
+              <img src="../../assets/user2.png" style="width: 35px; height: 35px;"> {{ sysUserName }}
+            </span>
+            <template #dropdown>
+              <el-dropdown-menu>
+                <el-dropdown-item @click="handleChangePassword">
+                  <el-icon><Edit /></el-icon>{{ $t("home.funpwd") }}
+                </el-dropdown-item>
+                <el-dropdown-item divided @click="logout">
+                  <el-icon><SwitchButton /></el-icon>{{ $t("home.funlogout") }}
+                </el-dropdown-item>
+              </el-dropdown-menu>
+            </template>
           </el-dropdown>
         </el-col>
         <el-col :span="2" class="language">
@@ -29,20 +31,34 @@
 
       <el-col :span="24" class="main">
         <aside>
-          <el-menu :default-active="$route.path" unique-opened @open="handleopen" @close="handleclose"
-                   @select="handleselect" background-color="#09446d"
-                   text-color="#D6E4E5" active-text-color="#fd6d21" router>
-            <template v-for="(item,index) in allMenus" v-if="!item.hidden">
-              <el-submenu :index="index+''" v-if="!item.leaf" :key="item.tag">
-                <template slot="title">
-                  <i :class="item.icon"/>{{ item.name }}
+          <el-menu 
+            :default-active="$route.path" 
+            unique-opened 
+            @open="handleopen" 
+            @close="handleclose"
+            @select="handleselect" 
+            background-color="#09446d"
+            text-color="#D6E4E5" 
+            active-text-color="#fd6d21" 
+            router>
+            <template v-for="(item,index) in allMenus" :key="item.tag">
+              <el-sub-menu :index="String(index)" v-if="!item.leaf && !item.hidden">
+                <template #title>
+                  <el-icon v-if="item.icon"><component :is="getIconComponent(item.icon)" /></el-icon>
+                  <span>{{ item.name }}</span>
                 </template>
-                <el-menu-item v-for="child in item.children" :index="child.path" v-if="!child.hidden" :key="child.tag">
-                  <i :class="child.icon"/>{{ child.name }}
+                <el-menu-item 
+                  v-for="child in item.children" 
+                  :index="child.path" 
+                  v-if="!child.hidden" 
+                  :key="child.tag">
+                  <el-icon v-if="child.icon"><component :is="getIconComponent(child.icon)" /></el-icon>
+                  <template #title>{{ child.name }}</template>
                 </el-menu-item>
-              </el-submenu>
-              <el-menu-item v-if="item.leaf" :index="item.path" :key="item.tag">
-                <i :class="item.icon"/>{{ item.name }}
+              </el-sub-menu>
+              <el-menu-item v-if="item.leaf && !item.hidden" :index="item.path" :key="item.tag">
+                <el-icon v-if="item.icon"><component :is="getIconComponent(item.icon)" /></el-icon>
+                <template #title>{{ item.name }}</template>
               </el-menu-item>
             </template>
           </el-menu>
@@ -50,7 +66,7 @@
         <section class="content-container">
           <div class="grid-content bg-purple-light">
             <el-col :span="24" class="Breadcrumb">
-             {{ $route.name }}
+              {{ $route.name }}
             </el-col>
             <el-col :span="24" class="content-wrapper">
               <transition>
@@ -66,9 +82,10 @@
         </section>
       </el-col>
     </el-row>
-    <el-dialog :title="$t('home.funpwd')" :visible.sync="ui.dialogVisible" width="650px" center class="addDio">
-      <el-form :model="editForm" :rules="editFormRules" label-width="190px" label-position="right" size="small"
-               ref="editForm">
+    
+    <el-dialog :title="$t('home.funpwd')" v-model="ui.dialogVisible" width="650px" center class="addDio">
+      <el-form :model="editForm" :rules="editFormRules" label-width="190px" label-position="right" size="default"
+               ref="editFormRef">
         <el-form-item :label="$t('home.pwdold')" prop="oldPassword">
           <el-input type="password" v-model="editForm.oldPassword" clearable class="edit-form-input"></el-input>
         </el-form-item>
@@ -79,218 +96,251 @@
           <el-input type="password" v-model="editForm.newPasswordAgain" clearable class="edit-form-input"></el-input>
         </el-form-item>
       </el-form>
-      <div slot="footer" class="dialog-footer">
-        <el-button @click.native="ui.dialogVisible = false">{{ $t("common.cancel") }}</el-button>
-        <el-button type="primary" @click.native="changePassword()">{{ $t("common.ok") }}</el-button>
-      </div>
+      <template #footer>
+        <div class="dialog-footer">
+          <el-button @click="ui.dialogVisible = false">{{ $t("common.cancel") }}</el-button>
+          <el-button type="primary" @click="changePassword()">{{ $t("common.ok") }}</el-button>
+        </div>
+      </template>
     </el-dialog>
   </div>
 </template>
 
-<script lang="ts">
-import Vue from "vue";
-import Component from "vue-class-component";
-import i18n from "../i18n";
-import * as   utils from "../../utils";
-import * as nav from "../../common/nav";
-import Language from "../common/Language.vue";
+<script setup lang="ts">
+import { ref, reactive, onMounted } from 'vue'
+import { useRoute, useRouter } from 'vue-router'
+import { useStore } from 'vuex'
+import { useI18n } from 'vue-i18n'
+import { ElMessage, ElMessageBox, type FormInstance } from 'element-plus'
+// import { Edit, SwitchButton, Home, User, Setting, Document } from '@element-plus/icons-vue'
+import * as utils from "../../utils"
+import * as nav from "../../common/nav"
 
-@Component({
-  components: {
-    'app-language': Language
-  }
+const { t } = useI18n()
+const route = useRoute()
+const router = useRouter()
+const store = useStore()
+
+// Refs
+const editFormRef = ref<FormInstance>()
+
+// Reactive data
+const allMenus = ref<any[]>([])
+const sysUserName = ref("")
+const sysUserAvatar = ref("")
+
+const form = reactive({
+  name: "",
+  region: "",
+  date1: "",
+  date2: "",
+  delivery: false,
+  type: [],
+  resource: "",
+  desc: ""
 })
 
-@Component
-export default class Home extends Vue {
-  allMenus = [];
-  sysUserName = "";
-  sysUserAvatar = "";
-  form = {
-    name: "",
-    region: "",
-    date1: "",
-    date2: "",
-    delivery: false,
-    type: [],
-    resource: "",
-    desc: ""
-  };
-  editForm = {} as any;
-  editFormRules = {
-    oldPassword: [
-      {required: true, message: i18n.t("home.oldisnull"), trigger: "blur"}
-    ],
-    newPassword: [
-      {required: true, message: i18n.t("home.newisnull"), trigger: "blur"}
-    ],
-    newPasswordAgain: [
-      {required: true, message: i18n.t("home.repeatnew"), trigger: "blur"},
-      {validator: this.validatePasswordAgain, trigger: "blur"}
-    ]
-  };
+const editForm = reactive<any>({})
 
-  ui = {
-    dialogVisible: false
-  };
+const editFormRules = reactive({
+  oldPassword: [
+    { required: true, message: t("home.oldisnull"), trigger: "blur" }
+  ],
+  newPassword: [
+    { required: true, message: t("home.newisnull"), trigger: "blur" }
+  ],
+  newPasswordAgain: [
+    { required: true, message: t("home.repeatnew"), trigger: "blur" },
+    { validator: validatePasswordAgain, trigger: "blur" }
+  ]
+})
 
-  validatePasswordAgain(rule: any, value: any, callback: any) {
-    if (value == "") {
-      callback(new Error(String(i18n.t("home.repeatnew"))));
-    } else if (value != this.editForm.newPassword) {
-      callback(new Error(String(i18n.t("home.notsame"))));
-    } else if (value == this.editForm.oldPassword) {
-      callback(new Error(String(i18n.t("home.issame"))));
-    } else {
-      callback();
-    }
+const ui = reactive({
+  dialogVisible: false
+})
+
+// Methods
+function validatePasswordAgain(rule: any, value: any, callback: any) {
+  if (value === "") {
+    callback(new Error(t("home.repeatnew")))
+  } else if (value !== editForm.newPassword) {
+    callback(new Error(t("home.notsame")))
+  } else if (value === editForm.oldPassword) {
+    callback(new Error(t("home.issame")))
+  } else {
+    callback()
   }
+}
 
-  handleopen() {
-    console.log('handleopen');
+const getIconComponent = (iconClass: string) => {
+  // Map Element UI icon classes to Element Plus icons
+  const iconMap: { [key: string]: any } = {
+    'el-icon-fa-home': Home,
+    'el-icon-fa-user': User,
+    'el-icon-fa-users': User,
+    'el-icon-fa-cogs': Setting,
+    'el-icon-s-custom': User,
+    'el-icon-s-data': Document,
+    'el-icon-postcard': Document,
   }
+  return iconMap[iconClass] || Document
+}
 
-  handleclose() {
-    console.log('handleclose');
-  }
+const handleopen = () => {
+  console.log('handleopen')
+}
 
-  handleselect(a: any, b: any) {
-  }
+const handleclose = () => {
+  console.log('handleclose')
+}
 
-  async handleChangePassword() {
-    let el: any = this.$refs.editForm;
-    if (el) {
-      el.resetFields();
-    }
-    this.editForm = {};
-    this.ui.dialogVisible = true;
-  }
+const handleselect = (a: any, b: any) => {
+  // Handle menu select
+}
 
-  async changePassword() {
-    let validated = await utils.validateForm(this.$refs.editForm);
-    if (!validated) return;
-    let confirmed = await utils.confirm(
-        this,
-        String(i18n.t("home.change-pwd-confirm")),
-        String(i18n.t("home.confirmTip"))
-    );
-    if (!confirmed) return;
-    let record = this.editForm;
-    let result = await utils.doPost(
-        this,
+const handleChangePassword = async () => {
+  editFormRef.value?.resetFields()
+  Object.assign(editForm, {})
+  ui.dialogVisible = true
+}
+
+const changePassword = async () => {
+  try {
+    await editFormRef.value?.validate()
+    
+    try {
+      await ElMessageBox.confirm(
+        t("home.change-pwd-confirm"),
+        t("home.confirmTip"),
+        { type: 'warning' }
+      )
+      
+      const record = editForm
+      const result = await utils.doPost(
+        { $router: router },
         "/api/sessions/passwordchange",
         record
-    );
-    if (!result.success) {
-      if (result.code = 302) {
-        utils.showWarning(String(i18n.t('user.incorrectpassword')));
+      )
+      
+      if (!result.success) {
+        if (result.code === 302) {
+          ElMessage.warning(t('user.incorrectpassword'))
+        } else {
+          ElMessage.warning(t('home.funpwdfal') + result.message)
+        }
       } else {
-        utils.showWarning(String(i18n.t('home.funpwdfal')) + result.message);
+        ElMessage.success(t("home.funpwdsuc"))
+        ui.dialogVisible = false
+        forceLogout()
       }
-    } else {
-      utils.showSuccess(String(i18n.t("home.funpwdsuc")));
-      this.ui.dialogVisible = false;
-      this.forceLogout();
+    } catch {
+      // User cancelled confirmation
     }
+  } catch {
+    // Form validation failed
   }
+}
 
-  async forceLogout() {
-    localStorage.removeItem("TOKEN");
-    localStorage.removeItem("user");
-    this.$router.push("/login");
+const forceLogout = async () => {
+  localStorage.removeItem("TOKEN")
+  localStorage.removeItem("user")
+  router.push("/login")
+}
+
+const logout = async () => {
+  try {
+    await ElMessageBox.confirm(
+      t("home.logoutmessage"),
+      t("home.confirmTip"),
+      { type: 'warning' }
+    )
+    
+    localStorage.removeItem("TOKEN")
+    localStorage.removeItem("user")
+    router.push("/login")
+  } catch {
+    // User cancelled
   }
+}
 
-  async logout() {
-    let confirmed = await utils.confirm(
-        this,
-        String(i18n.t("home.logoutmessage")),
-        String(i18n.t("home.confirmTip"))
-    );
-    if (confirmed) {
-      localStorage.removeItem("TOKEN");
-      localStorage.removeItem("user");
-      this.$router.push("/login");
-    }
+const loadAllMenus = async () => {
+  const result = await utils.doGet({ $router: router }, "/api/sessions/permissions")
+  if (result.success) {
+    allMenus.value = result.data
   }
-
-  async loadAllMenus() {
-    let result = await utils.doGet(this, "/api/sessions/permissions");
-    if (result.success) {
-      this.allMenus = result.data;
-    }
-    var availButtons = [];
-    for (var i = 0; i < this.allMenus.length; i++) {
-      var menu: any = this.allMenus[i];
-      if (menu.children) {
-        var buttons = menu.children;
-        for (var k = 0; k < buttons.lenght; k++)
-          availButtons.push(buttons[k].tag);
+  
+  const availButtons: string[] = []
+  for (let i = 0; i < allMenus.value.length; i++) {
+    const menu: any = allMenus.value[i]
+    if (menu.children) {
+      const buttons = menu.children
+      for (let k = 0; k < buttons.length; k++) {
+        availButtons.push(buttons[k].tag)
       }
-      if (menu.children) {
-        for (var j = 0; j < menu.children.length; j++) {
-          var child = menu.children[j];
-          if (child.children) {
-            var buttons = child.children;
-            for (var k = 0; k < buttons.length; k++)
-              availButtons.push(buttons[k].tag);
+    }
+    if (menu.children) {
+      for (let j = 0; j < menu.children.length; j++) {
+        const child = menu.children[j]
+        if (child.children) {
+          const buttons = child.children
+          for (let k = 0; k < buttons.length; k++) {
+            availButtons.push(buttons[k].tag)
           }
         }
       }
     }
-    this.$store.commit("setButtons", {
-      buttons: new Set(availButtons)
-    });
   }
+  
+  store.commit("setButtons", {
+    buttons: new Set(availButtons)
+  })
+}
 
-  async availButtons() {
-    let result = await utils.doGet(this, "/api/sessions/permission/tag");
-    let availableTags = [];
-    if (result.success) {
-      availableTags = result.data;
-      let menus: any = nav.getNavData();
-      this.hiddenMenus(menus, availableTags);
-      this.allMenus = menus;
-    }
-    this.$store.commit("setButtons", {
-      buttons: new Set(availableTags)
-    });
+const availButtons = async () => {
+  const result = await utils.doGet({ $router: router }, "/api/sessions/permission/tag")
+  let availableTags: string[] = []
+  
+  if (result.success) {
+    availableTags = result.data
+    const menus: any = nav.getNavData()
+    hiddenMenus(menus, availableTags)
+    allMenus.value = menus
   }
+  
+  store.commit("setButtons", {
+    buttons: new Set(availableTags)
+  })
+}
 
-  hiddenMenus(menus: any, tags: any) {
-    for (let menu of menus) {
-      if (menu.whiteList) {
-        menu.hidden == false;
-      } else {
-        menu.hidden = !this.compareTag(menu.tag, tags);
-      }
-      if (menu.children) {
-        this.hiddenMenus(menu.children, tags);
-      }
+const hiddenMenus = (menus: any[], tags: string[]) => {
+  for (const menu of menus) {
+    if (menu.whiteList) {
+      menu.hidden = false
+    } else {
+      menu.hidden = !compareTag(menu.tag, tags)
     }
-  }
-
-  compareTag(tag: string, tags: Array<string>) {
-    let flag = false;
-    for (let t of tags) {
-      if (t == tag) {
-        flag = true;
-        break;
-      }
+    if (menu.children) {
+      hiddenMenus(menu.children, tags)
     }
-    return flag;
-  }
-
-  mounted() {
-    let userString = sessionStorage.getItem("user");
-    if (userString) {
-      let user = JSON.parse(userString);
-      this.sysUserName = user.realname || user.userName;
-      this.sysUserAvatar = user.avatar || "";
-    }
-    this.availButtons();
   }
 }
+
+const compareTag = (tag: string, tags: string[]) => {
+  return tags.includes(tag)
+}
+
+// Lifecycle
+onMounted(() => {
+  const userString = sessionStorage.getItem("user")
+  if (userString) {
+    const user = JSON.parse(userString)
+    sysUserName.value = user.realname || user.userName
+    sysUserAvatar.value = user.avatar || ""
+  }
+  availButtons()
+})
 </script>
+
 <style scoped>
 .container {
   position: absolute;
@@ -356,7 +406,6 @@ export default class Home extends Vue {
 }
 
 .container .content-container {
-  /*min-width: 1200px;*/
   background: #f3f5f7;
   position: absolute;
   right: 0px;
